# src/common/swagger/swagger-output.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "3.0.0",
        "title": "allinjob API Docs",
        "description": "올인잡 api 문서입니다"
    },
    "host": "https://allinjob.co.kr",
    "servers": [
        {
            "url": "http://localhost:4000"
        },
        {
            "url": "https://allinjob.co.kr"
        }
    ],
    "basePath": "/",
...
    "schemes": ["http", "https"],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "CookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "refreshToken"
            }
        }
    },
    "paths": {
        "/user/findOneUserByEmail": {
            "get": {
                "tags": ["Users"],
                "summary": "회원가입한 유저 이메일 찾아오기",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "email",
                            "example": "abc@abc.com"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "응답 성공",
                        "content": {
                            "application/json": {
                                "example": {
                                    "value": {
                                        "data": {
                                            "id": "adfhttwerqwrasxczx",
                                            "email": "abc@abc.com",
                                            "provider": "kakao, naver, google",
                                            "name": "홍길동",
                                            "nickname": "닉네임",
                                            "phone": "01012345678",
                                            "profileImage": "awihqwifbjqjfbo",
                                            "major": "컴퓨터공학",
                                            "interests": [
                                                {
                                                    "id": "asdsadasdasdasdasdsef",
                                                    "userId": "adfhttwerqwrasxczx",
                                                    "interestId": "adwwafefewfsdfdsgew",
                                                    "keywordId": "asdwfwgregrerfsfsefe",
                                                    "interest": {
                                                        "id": "adwwafefewfsdfdsgew",
                                                        "interest": "공모전, 대외활동, 자격증, 어학, 인턴"
                                                    },
                                                    "keyword": {
                                                        "id": "asdwfwgregrerfsfsefe",
                                                        "keyword": "기획/아이디어, 광고/마케팅, 사진/영상/UCC, 디자인/순수예술/공예, 네이밍/슬로건, 캐릭터/만화/게임, 건축/건설/인테리어, 과학/공학, 예체능/패션, 전시/패션, 문학/시나리오, 해외, 학술, 창업, 기타, 서포터즈, 해외탐방/무료, 해외탐방/유료, 봉사단/해외, 봉사단/국내, 마케터, 기자단, 강연, 멘토링, 경영/회계/사무, 보건/의료, 문화예술/다지안/방송, 사회복지/종교, 이용/숙박/여행/오락/스포츠, 음식서비스, 기계, 정보통신, 식품/가공, 환경/에너지, TOEIC, TOEIC.speaking and writing tests, TOEFL, TOEIC bridge, TOEIC bridge.speaking and writing tests, 대기업, 중소기업, 공공기관/공기업, 외국계기업, 중견기업, 비영리단체/협회/재단, 동아리/학생자치단체"
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                },
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "provider": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "nickname": {
                                                    "type": "string"
                                                },
                                                "phone": {
                                                    "type": "string"
                                                },
                                                "profileImage": {
                                                    "type": "string"
                                                },
                                                "major": {
                                                    "type": "string"
                                                },
                                                "interests": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "id": {
                                                                "type": "string"
                                                            },
                                                            "userId": {
                                                                "type": "string"
                                                            },
                                                            "interestId": {
                                                                "type": "string"
                                                            },
                                                            "keywordId": {
                                                                "type": "string"
                                                            },
                                                            "interest": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "string"
                                                                    },
                                                                    "interest": {
                                                                        "type": "string"
                                                                    }
                                                                }
                                                            },
                                                            "keyword": {
                                                                "type": "object",
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "string"
                                                                    },
                                                                    "keyword": {
                                                                        "type": "string"
                                                                    }
                                                                }
                                                            }
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "요청 형식 에러",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "email must be an email"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "회원가입한 유저가 없는 경우",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "회원가입한 유저가 없습니다"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 에러",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "서버 문제"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/findOneUserByID": {
            "get": {
                "tags": ["Users"],
                "summary": "아이디 찾기",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "홍길동"
                        },
                        "description": "이름"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "01012345678"
                        },
                        "description": "전화번호"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "응답 성공",
                        "content": {
                            "application/json": {
                                "example": {
                                    "data": {
                                        "email": "abc@abc.com",
                                        "provider": "kakao, naver, google"
                                    }
                                },
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "email": {
                                                    "type": "string"
                                                },
                                                "provider": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "요청 형식 에러",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": [
                                        "name : name must be a string",
                                        "phone : phone must be a string"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "일치하는 아이디가 없는 경우",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "일치하는 아이디가 없습니다"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 에러",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "서버 문제"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/isNickname": {
            "get": {
                "tags": ["Users"],
                "summary": "닉네임 중복 여부 확인하기",
                "parameters": [
                    {
                        "name": "nickname",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "닉네임"
                        },
                        "description": "닉네임"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "닉네임 사용 가능",
                        "content": {
                            "application/json": {
                                "example": {
                                    "data": true
                                },
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "examples": {
                                    "NoLogin": {
                                        "summary": "닉네임 중복되는 경우",
                                        "value": {
                                            "error": "이미 사용중인 닉네임입니다"
                                        }
                                    },
                                    "TypeError": {
                                        "summary": "닉네임 형식이 잘못된 경우",
                                        "value": {
                                            "error": "nickname: nickname must match /^[a-zA-Z가-힣]+$/ regular expression"
                                        }
                                    }
                                },
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "string"
                                        },
                                        {
                                            "type": "string"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 에러",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "서버 문제"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/user/sendTokenSMS": {
            "post": {
                "tags": ["Users"],
                "summary": "문자 인증번호 요청하기",
                "description": "",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "01012345678"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SMS 토큰 전송",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "SMS 토큰 전송 성공": {
                                        "value": {
                                            "data": true
                                        }
                                    },
                                    "SMS 토큰 전송 실패": {
                                        "value": {
                                            "data": false
                                        }
                                    }
                                },
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "true": {
                                                    "type": "boolean"
                                                }
                                            }
                                        },
                                        {
                                            "properties": {
                                                "false": {
                                                    "type": "boolean"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "요청 형식 에러",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": ["phone : phone must be a string"]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 에러",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "서버 문제"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/validateToken": {
            "post": {
                "tags": ["Users"],
                "summary": "문자 인증번호 검증하기",
                "description": "",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "token": {
                                        "type": "number",
                                        "example": 11111
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "01012345678"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SMS 토큰 검증",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "SMS 토큰 검증 성공": {
                                        "value": {
                                            "data": true
                                        }
                                    },
                                    "SMS 토큰 검증 실패": {
                                        "value": {
                                            "data": false
                                        }
                                    }
                                },
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "true": {
                                                    "type": "boolean"
                                                }
                                            }
                                        },
                                        {
                                            "properties": {
                                                "false": {
                                                    "type": "boolean"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "요청 형식 에러",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": [
                                        "token : token must be a number",
                                        "phone : phone must be a string"
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 에러",
                        "content": {
                            "application/json": {
                                "example": {
                                    "error": "서버 문제"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}